refactor: revert Proposals component and add support for lazer option
