+++ governance/xc_admin/packages/xc_admin_frontend/hooks/useMultisig.ts
@@ -133,7 +133,7 @@ export const useMultisig = (): MultisigHookData => {
     }
 
     return { cancel, fetchData }
-  }, [multisigCluster, urlsIndex, connection])
+  }, [readOnlySquads, multisigCluster, urlsIndex])
 
   useEffect(() => {
     const { cancel, fetchData } = refreshData()
